pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-trading-system"
version = "1.0.0"
description = "简化量化交易系统 - 集成币安数据、Alpaca交易、backtesting.py回测"
authors = [
    {name = "Trading System Team", email = "team@trading.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # 数据获取
    "requests>=2.31.0",
    
    # 回测框架
    "backtesting>=0.3.3",
    
    # 数据处理
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # 技术指标
    "TA-Lib>=0.4.0",
    
    # 交易API
    "alpaca-trade-api>=3.0.0",
    
    # 环境变量
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/trading-team/simple-trading-system"
Repository = "https://github.com/trading-team/simple-trading-system.git"
Issues = "https://github.com/trading-team/simple-trading-system/issues"

[project.scripts]
trading-system = "simple_trading_system.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["simple_trading_system*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=simple_trading_system --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true